Extracted AskWindow from AskButton
